+++ internal/css_lexer/css_lexer_test.go
@@ -7,13 +7,14 @@ import (
 	"github.com/evanw/esbuild/internal/test"
 )
 
-func lexToken(contents string) T {
+func lexToken(contents string) (T, string) {
 	log := logger.NewDeferLog()
 	tokens := Tokenize(log, test.SourceForTest(contents))
 	if len(tokens) > 0 {
-		return tokens[0].Kind
+		t := tokens[0]
+		return t.Kind, t.DecodedText(contents)
 	}
-	return TEndOfFile
+	return TEndOfFile, ""
 }
 
 func lexerError(contents string) string {
@@ -75,41 +76,50 @@ func TestTokens(t *testing.T) {
 		contents := it.contents
 		token := it.token
 		t.Run(contents, func(t *testing.T) {
-			test.AssertEqual(t, lexToken(contents), token)
+			kind, _ := lexToken(contents)
+			test.AssertEqual(t, kind, token)
 		})
 	}
 }
 
 func TestStringParsing(t *testing.T) {
-	test.AssertEqual(t, ContentsOfStringToken("\"foo\""), "foo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\oo\""), "foo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\\"o\""), "f\"o")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\\\o\""), "f\\o")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\\no\""), "fo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\\ro\""), "fo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\\r\no\""), "fo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\\fo\""), "fo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\6fo\""), "foo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\6f o\""), "foo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\6f  o\""), "fo o")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\fffffffo\""), "f\uFFFDfo")
-	test.AssertEqual(t, ContentsOfStringToken("\"f\\10abcdeo\""), "f\U0010ABCDeo")
+	contentsOfStringToken := func(contents string) string {
+		t.Helper()
+		kind, text := lexToken(contents)
+		test.AssertEqual(t, kind, TString)
+		return text
+	}
+	test.AssertEqual(t, contentsOfStringToken("\"foo\""), "foo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\oo\""), "foo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\\"o\""), "f\"o")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\\\o\""), "f\\o")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\\no\""), "fo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\\ro\""), "fo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\\r\no\""), "fo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\\fo\""), "fo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\6fo\""), "foo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\6f o\""), "foo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\6f  o\""), "fo o")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\fffffffo\""), "f\uFFFDfo")
+	test.AssertEqual(t, contentsOfStringToken("\"f\\10abcdeo\""), "f\U0010ABCDeo")
 }
 
 func TestURLParsing(t *testing.T) {
-	contentsOfURLToken := func(raw string) string {
-		text, _ := ContentsOfURLToken(raw)
+	contentsOfURLToken := func(expected T, contents string) string {
+		t.Helper()
+		kind, text := lexToken(contents)
+		test.AssertEqual(t, kind, expected)
 		return text
 	}
-	test.AssertEqual(t, contentsOfURLToken("url(foo)"), "foo")
-	test.AssertEqual(t, contentsOfURLToken("url(  foo\t\t)"), "foo")
-	test.AssertEqual(t, contentsOfURLToken("url(f\\oo)"), "foo")
-	test.AssertEqual(t, contentsOfURLToken("url(f\\\"o)"), "f\"o")
-	test.AssertEqual(t, contentsOfURLToken("url(f\\'o)"), "f'o")
-	test.AssertEqual(t, contentsOfURLToken("url(f\\)o)"), "f)o")
-	test.AssertEqual(t, contentsOfURLToken("url(f\\6fo)"), "foo")
-	test.AssertEqual(t, contentsOfURLToken("url(f\\6f o)"), "foo")
-	test.AssertEqual(t, contentsOfURLToken("url(f\\6f  o)"), "fo o")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(foo)"), "foo")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(  foo\t\t)"), "foo")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(f\\oo)"), "foo")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(f\\\"o)"), "f\"o")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(f\\'o)"), "f'o")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(f\\)o)"), "f)o")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(f\\6fo)"), "foo")
+	test.AssertEqual(t, contentsOfURLToken(TURL, "url(f\\6f o)"), "foo")
+	test.AssertEqual(t, contentsOfURLToken(TBadURL, "url(f\\6f  o)"), "url(f\\6f  o)")
 }
 
 func TestComment(t *testing.T) {